* mixer-new-fifo-arch:
  ProjectRenderer: lock Mixer while calling Song::{start,stop}Export()
  ProjectRenderer: start thread with normal priority
  ProjectRenderer: renamed OutputSettings to EncoderSettings + Doxygen comments
  Mixer: rewrote processing chain of rendered audio buffers

+++ include/cpuload_widget.h
@@ -51,7 +51,7 @@ protected slots:
 
 
 private:
-	Uint8 m_currentLoad;
+	int m_currentLoad;
 
 	QPixmap m_temp;
 	QPixmap m_background;
